Extracted `TestColumnStatsIndex`
